requirements.txt
pandas
aiogram
telebot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
